TAsk services updated updated method

+++ backend/taskmanager/src/main/java/com/example/taskmanager/service/TaskServices.java
@@ -105,16 +105,47 @@ public class TaskServices {
     }
 
     public TaskResponse gerTaskById(Long id){
+        User user= getCurrentUser();
+
         Task task =taskRepository.findById(id)
                 .orElseThrow(()-> new TaskNotFoundException("Task Not Found" + id));
+
+        if(!task.getUser().getId().equals(user.getId())){
+            throw new RuntimeException("nahi hai koi user login kar ya register kar ");
+        }
         return taskMapper.toResponse(task);
     }
 
-    public TaskResponse updateTask(Long id , Task updatedTask){
+    public TaskResponse updateTask(Long id , Task updatedTask , TaskRequest request){
+
+        User user= getCurrentUser();
+
+
         Task existingTask= taskRepository.findById(id)
                 .orElseThrow(()-> new TaskNotFoundException("Cannot delete it" + id ));
 
-        existingTask.setTitle(updatedTask.getTitle());
+        if(!existingTask.getUser().getId().equals(user.getId())){
+            throw new RuntimeException("Acces denied");
+        }
+        existingTask.setTitle(request.getTitle());
+        existingTask.setDescription(request.getDescription());
+
+        if(request.getPriority()!=null){
+            existingTask.setTaskPriority(TaskPriority.valueOf(request.getPriority().toUpperCase()));
+        }
+        if (request.getStatus()!=null){
+            existingTask.setTaskStatus(
+                    TaskStatus.valueOf(request.getStatus().toUpperCase())
+            );
+        }
+
+        existingTask.setDueDate(request.getDueDate());
+        existingTask.setCategory(request.getCategory());
+        existingTask.setEstimatedMinutes(request.getEstimatedMinutes());
+
+
+        existingTask.setUpdatedAt(LocalDateTime.now());
+
 //        existingTask.setCompleted(updatedTask.isCompleted());
 
         Task updatedTasks=taskRepository.save(existingTask);
@@ -124,10 +155,21 @@ public class TaskServices {
         }
 
     public void deleteTask(Long id){
+        User user= getCurrentUser();
         Task task = taskRepository.findById(id).orElseThrow(()-> new TaskNotFoundException("Task Not FOund" + id));
+        if(!task.getUser().getId().equals(user.getId())){
+            throw new RuntimeException("Acces denied");
+        }
         taskRepository.delete(task);
     }
 
+    private User getCurrentUser(){
+        Authentication authentication=SecurityContextHolder.getContext().getAuthentication();
+        String username = authentication.getName();
+
+        return userRepository.findByName(username)
+                .orElseThrow(()->new RuntimeException("No user found"));
+    }
 
     public List<TaskResponse> getTasksByUser(Long userId){
         return taskRepository.findByUserId(userId)
